requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
pdfplumber
aiofiles
astrapy
